+++ Assets/Sources/Weapons/Bullet/Bullet.cs
@@ -7,21 +7,21 @@ public class Bullet : MonoBehaviour
     private CharacterController characterController;
     private Vector3 velocity;
     private float speed;
-    private float damage;
+    private int damage;
     private int ignoreCollisionBulletLayer = 3;
     private int timeOfLife = 2;
 
-	public void StartFire(TurretController owner, Vector3 target, float speed, float damage)
+	public void StartFire(TurretController owner, Vector3 target, float speed, int damage)
 	{
         StartFire(owner.transform, target, speed, damage);
 	}
 
-	public void StartFire(ActorController owner, Vector3 target, float speed, float damage)
+	public void StartFire(ActorController owner, Vector3 target, float speed, int damage)
     {
 		StartFire(owner.transform, target, speed, damage);
 	}
 
-    private void StartFire(Transform owner, Vector3 target, float speed, float damage)
+    private void StartFire(Transform owner, Vector3 target, float speed, int damage)
     {
 		this.gameObject.layer = ignoreCollisionBulletLayer;
 		characterController = GetComponent<CharacterController>();
